LPS-170680 Extract Settings card to its own component
